refactor(assets): extract upsert into lib/asset-upsert.ts; asset-service is now a barrel
Final asset-service.ts split PR. Pure relocation of the source-keyed dedup spine
out of asset-service.ts into asset-upsert.ts: findBySourcePath, upsertFromSource,
resolveStoreFile, and the sha256File/findSameTaskContentMatch/upsertFromSourceInner
internals. asset-upsert imports the create/read core from asset-core and the
version mutations from asset-mutations (never the barrel), so no cycle.

asset-service.ts is now a 50-line pure re-export barrel over asset-core +
asset-mutations + asset-upsert — the 21 consumers + tests still import every
symbol from `…/lib/asset-service` unchanged. Function bodies byte-for-byte
unchanged across the whole 3-PR split (681 → barrel + 3 cohesive modules).

Verified: typecheck/eslint/assets suite 242-0/full suite 5124-0/binary build
(stamps reverted) + boot smoke (assets activates; bakin_assets reindex clean).
